+++ docs/changelog.md
@@ -29,6 +29,7 @@ category: Administration
 * refactor: do not create empty project specific settings when not needed
 * test: improve test coverage and code quality for project configuration action
 * test: add test coverage for project configuration service
+* test: verify correct overwriting of project specific configuration
 
 ### [20.10.0] - 2020-10-08